+++ src/test/java/com/bitmart/examples/futures/trade/GetOrderHistory.java
@@ -5,12 +5,12 @@ import com.bitmart.api.CloudContext;
 import com.bitmart.api.common.CloudException;
 import com.bitmart.api.common.CloudResponse;
 import com.bitmart.api.key.CloudKey;
-import com.bitmart.api.request.contract.prv.GetTransactionHistoryRequest;
+import com.bitmart.api.request.contract.prv.GetOrderHistoryRequest;
 import com.bitmart.examples.Example;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class GetTransactionHistory {
+public class GetOrderHistory {
 
     private static final String API_KEY = Example.YOUR_API_KEY;
     private static final String API_SECRET = Example.YOUR_API_SECRET;
@@ -20,12 +20,21 @@ public class GetTransactionHistory {
         Call call = new Call(new CloudContext(Example.FUTURES_HOST, new CloudKey(API_KEY, API_SECRET, API_MEMO)));
 
         try {
-            final CloudResponse cloudResponse = call.callCloud(new GetTransactionHistoryRequest()
+            final CloudResponse cloudResponse = call.callCloud(new GetOrderHistoryRequest()
                     .setSymbol("BTCUSDT")
-                    .setFlowType(0) // All
-                    .setStartTime(System.currentTimeMillis() - 86400000L) // 1 day ago
-                    .setEndTime(System.currentTimeMillis()) // now
-                    .setPageSize(100)
+                    .setAccount("futures") // futures account
+                    .setOrderId("123456789")
+            );
+            System.out.println(cloudResponse);
+        } catch (CloudException e) {
+            log.error("Error response: " + e.getMessage());
+        }
+
+        try {
+            final CloudResponse cloudResponse = call.callCloud(new GetOrderHistoryRequest()
+                    .setSymbol("BTCUSDT")
+                    .setStartTime(System.currentTimeMillis()/1000 - 86400L) // 1 day ago
+                    .setEndTime(System.currentTimeMillis()/1000) // now
                     .setAccount("futures") // futures account
             );
             System.out.println(cloudResponse);