+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 Changelog
 =========================
 
+### v2.4.0 Release
+#### New Features
+- New
+  - `/contract/private/modify-limit-order` Modify Limit Orders(SIGNED)
+  - `/contract/private/cancel-all-after` Timed Cancel All Orders (SIGNED)
+  - `/contract/private/set-position-mode` Set Position Mode (SIGNED)
+  - `/contract/private/get-position-mode` Get Position Mode (KEYED)
+  - `/contract/private/position-v2` Get Current Position V2 (KEYED)
+  - `/contract/public/leverage-bracket` Get Current Leverage Risk Limit
+  - `/contract/public/market-trade` Query the latest trade data
+  - `/contract/public/markprice-kline` Get MarkPrice K-line
+- Update
+  - `/contract/private/submit-order` Submit Order(SIGNED): Add new request field **stp_mode**
+  - `/contract/private/order` Get Order Detail (KEYED): Add new request field **account**
+  - `/contract/private/order-history` Get Order History (KEYED): Add new request field **account**, **order_id**, **client_order_id**
+  - `/contract/private/position` Get Current Position (KEYED): Add new request field **account**
+  - `/contract/private/position-risk` Get Current Position Risk Details(KEYED): Add new request field **account**
+  - `/contract/private/trades` Get Order Trade (KEYED): Add new request field **account**
+  - `/contract/private/transaction-history` Get Transaction History (KEYED): Add new request field **account**
+  - `/contract/private/trades` Get Order Trade: Request field **symbol** changed to optional
+#### Improvements
+- Upgrade **commons-lang3** to version **3.18.0**
+- Upgrade **commons-codec** to version **1.16.0**
+ 
+
+---
+
 
 ### v2.3.0 Release
 #### New Features